new version and fix conflicts
